working on making codecs extensible
