+++ zarr/v3/abc/codec.py
@@ -11,20 +11,19 @@
 from __future__ import annotations
 
 from abc import abstractmethod, ABC
-from typing import TYPE_CHECKING, Optional
+from typing import TYPE_CHECKING, Optional, Type
 
 import numpy as np
 
-from zarr.v3.common import BytesLike
+from zarr.v3.common import BytesLike, SliceSelection
+from zarr.v3.store import StorePath
 
 
 if TYPE_CHECKING:
-    from zarr.v3.metadata import CoreArrayMetadata
+    from zarr.v3.metadata import CoreArrayMetadata, CodecMetadata
 
 
 class Codec(ABC):
-    supports_partial_decode: bool
-    supports_partial_encode: bool
     is_fixed_size: bool
     array_metadata: CoreArrayMetadata
 
@@ -34,6 +33,12 @@ class Codec(ABC):
 
     def resolve_metadata(self) -> CoreArrayMetadata:
         return self.array_metadata
+
+    @classmethod
+    def from_metadata(
+        cls, codec_metadata: "CodecMetadata", array_metadata: CoreArrayMetadata
+    ) -> "Type[Codec]":
+        pass
 
 
 class ArrayArrayCodec(Codec):
@@ -65,6 +70,27 @@ class ArrayBytesCodec(Codec):
         self,
         chunk_array: np.ndarray,
     ) -> Optional[BytesLike]:
+        pass
+
+
+class ArrayBytesCodecPartialDecodeMixin:
+    @abstractmethod
+    async def decode_partial(
+        self,
+        store_path: StorePath,
+        selection: SliceSelection,
+    ) -> Optional[np.ndarray]:
+        pass
+
+
+class ArrayBytesCodecPartialEncodeMixin:
+    @abstractmethod
+    async def encode_partial(
+        self,
+        store_path: StorePath,
+        chunk_array: np.ndarray,
+        selection: SliceSelection,
+    ) -> None:
         pass
