+++ zarr/v3/codecs/crc32c_.py
@@ -1,0 +1,57 @@
+from __future__ import annotations
+
+from typing import (
+    TYPE_CHECKING,
+    Literal,
+    Optional,
+)
+
+import numpy as np
+from attr import frozen, field
+from crc32c import crc32c
+
+from zarr.v3.abc.codec import BytesBytesCodec
+from zarr.v3.codecs.registry import register_codec
+from zarr.v3.common import BytesLike
+
+if TYPE_CHECKING:
+    from zarr.v3.metadata import CoreArrayMetadata
+
+
+@frozen
+class Crc32cCodecMetadata:
+    name: Literal["crc32c"] = field(default="crc32c", init=False)
+
+
+@frozen
+class Crc32cCodec(BytesBytesCodec):
+    array_metadata: CoreArrayMetadata
+    is_fixed_size = True
+
+    @classmethod
+    def from_metadata(
+        cls, codec_metadata: Crc32cCodecMetadata, array_metadata: CoreArrayMetadata
+    ) -> Crc32cCodec:
+        return cls(array_metadata=array_metadata)
+
+    async def decode(
+        self,
+        chunk_bytes: bytes,
+    ) -> BytesLike:
+        crc32_bytes = chunk_bytes[-4:]
+        inner_bytes = chunk_bytes[:-4]
+
+        assert np.uint32(crc32c(inner_bytes)).tobytes() == bytes(crc32_bytes)
+        return inner_bytes
+
+    async def encode(
+        self,
+        chunk_bytes: bytes,
+    ) -> Optional[BytesLike]:
+        return chunk_bytes + np.uint32(crc32c(chunk_bytes)).tobytes()
+
+    def compute_encoded_size(self, input_byte_length: int) -> int:
+        return input_byte_length + 4
+
+
+register_codec("crc32c", Crc32cCodec, Crc32cCodecMetadata)
