+++ zarr/v3/codecs/zstd.py
@@ -1,0 +1,73 @@
+from __future__ import annotations
+
+from typing import (
+    TYPE_CHECKING,
+    Literal,
+    Optional,
+)
+
+from attr import frozen, field
+from zstandard import ZstdCompressor, ZstdDecompressor
+
+from zarr.v3.abc.codec import BytesBytesCodec
+from zarr.v3.codecs.registry import register_codec
+from zarr.v3.common import BytesLike, to_thread
+
+if TYPE_CHECKING:
+    from zarr.v3.metadata import CoreArrayMetadata
+
+
+@frozen
+class ZstdCodecConfigurationMetadata:
+    level: int = 0
+    checksum: bool = False
+
+
+@frozen
+class ZstdCodecMetadata:
+    configuration: ZstdCodecConfigurationMetadata
+    name: Literal["zstd"] = field(default="zstd", init=False)
+
+
+@frozen
+class ZstdCodec(BytesBytesCodec):
+    array_metadata: CoreArrayMetadata
+    configuration: ZstdCodecConfigurationMetadata
+    is_fixed_size = True
+
+    @classmethod
+    def from_metadata(
+        cls, codec_metadata: ZstdCodecMetadata, array_metadata: CoreArrayMetadata
+    ) -> ZstdCodec:
+        return cls(
+            array_metadata=array_metadata,
+            configuration=codec_metadata.configuration,
+        )
+
+    def _compress(self, data: bytes) -> bytes:
+        ctx = ZstdCompressor(
+            level=self.configuration.level, write_checksum=self.configuration.checksum
+        )
+        return ctx.compress(data)
+
+    def _decompress(self, data: bytes) -> bytes:
+        ctx = ZstdDecompressor()
+        return ctx.decompress(data)
+
+    async def decode(
+        self,
+        chunk_bytes: bytes,
+    ) -> BytesLike:
+        return await to_thread(self._decompress, chunk_bytes)
+
+    async def encode(
+        self,
+        chunk_bytes: bytes,
+    ) -> Optional[BytesLike]:
+        return await to_thread(self._compress, chunk_bytes)
+
+    def compute_encoded_size(self, _input_byte_length: int) -> int:
+        raise NotImplementedError
+
+
+register_codec("zstd", ZstdCodec, ZstdCodecMetadata)
