+++ zarr/v3/metadata.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import json
 from asyncio import AbstractEventLoop
 from enum import Enum
-from typing import Any, Dict, List, Literal, Optional, Tuple, Union
+from typing import Any, Dict, List, Literal, Optional, Protocol, Tuple, Union
 
 import numpy as np
 from attr import asdict, field, frozen
@@ -142,103 +142,9 @@ class V2ChunkKeyEncodingMetadata:
 ChunkKeyEncodingMetadata = Union[DefaultChunkKeyEncodingMetadata, V2ChunkKeyEncodingMetadata]
 
 
-BloscShuffle = Literal["noshuffle", "shuffle", "bitshuffle"]
-
-
-@frozen
-class BloscCodecConfigurationMetadata:
-    typesize: int
-    cname: Literal["lz4", "lz4hc", "blosclz", "zstd", "snappy", "zlib"] = "zstd"
-    clevel: int = 5
-    shuffle: BloscShuffle = "noshuffle"
-    blocksize: int = 0
-
-
-blosc_shuffle_int_to_str: Dict[int, BloscShuffle] = {
-    0: "noshuffle",
-    1: "shuffle",
-    2: "bitshuffle",
-}
-
-
-@frozen
-class BloscCodecMetadata:
-    configuration: BloscCodecConfigurationMetadata
-    name: Literal["blosc"] = "blosc"
-
-
-@frozen
-class BytesCodecConfigurationMetadata:
-    endian: Optional[Literal["big", "little"]] = "little"
-
-
-@frozen
-class BytesCodecMetadata:
-    configuration: BytesCodecConfigurationMetadata
-    name: Literal["bytes"] = "bytes"
-
-
-@frozen
-class TransposeCodecConfigurationMetadata:
-    order: Union[Literal["C", "F"], Tuple[int, ...]] = "C"
-
-
-@frozen
-class TransposeCodecMetadata:
-    configuration: TransposeCodecConfigurationMetadata
-    name: Literal["transpose"] = "transpose"
-
-
-@frozen
-class GzipCodecConfigurationMetadata:
-    level: int = 5
-
-
-@frozen
-class GzipCodecMetadata:
-    configuration: GzipCodecConfigurationMetadata
-    name: Literal["gzip"] = "gzip"
-
-
-@frozen
-class ZstdCodecConfigurationMetadata:
-    level: int = 0
-    checksum: bool = False
-
-
-@frozen
-class ZstdCodecMetadata:
-    configuration: ZstdCodecConfigurationMetadata
-    name: Literal["zstd"] = "zstd"
-
-
-@frozen
-class Crc32cCodecMetadata:
-    name: Literal["crc32c"] = "crc32c"
-
-
-@frozen
-class ShardingCodecConfigurationMetadata:
-    chunk_shape: ChunkCoords
-    codecs: List["CodecMetadata"]
-    index_codecs: List["CodecMetadata"]
-
-
-@frozen
-class ShardingCodecMetadata:
-    configuration: ShardingCodecConfigurationMetadata
-    name: Literal["sharding_indexed"] = "sharding_indexed"
-
-
-CodecMetadata = Union[
-    BloscCodecMetadata,
-    BytesCodecMetadata,
-    TransposeCodecMetadata,
-    GzipCodecMetadata,
-    ZstdCodecMetadata,
-    ShardingCodecMetadata,
-    Crc32cCodecMetadata,
-]
+class CodecMetadata(Protocol):
+    name: str
+    configuration: Optional[Any]
 
 
 @frozen
